requirements.txt
fastapi
uvicorn
google-genai
pytesseract
pillow
python-dotenv